feat(integration): 1만 가구 E2E 스트리밍 연동 및 15분 단위 렌더링 규격 동기화

+++ energy-api/src/main/java/com/energy/api/dto/AppDto.java
@@ -15,7 +15,10 @@ public class AppDto {
         private Double monthlyTarget;
         private Double monthlyUsed;
         private Double savingPercent;
-        private Integer monthlySaving;
+
+        // 수정 전: private Integer monthlySaving;
+        private Double monthlySaving; // 수정 후: 전력 도메인 정밀도 확보 및 빌드 에러 방어
+
         private Double co2Reduction;
         private Integer totalPoints;
         private List<Double> hourlyActual;
